Completed except for page-filling main
Matched the excercise style reasonably closely, except for the main
element's height.

I tried setting the html and body height to 100%, but that alone gives
me a small scroll-bar (even if that is the only change I make).
Ignoring that problem I don't know how I can tell the main element to
take up the remaining height. We've learned to do this with the 1 single
top element by giving it a min-height: 100% but doing that on main gives
us a large scrollbar because there is the header on top of it. So main's
height should actually be 100% - height(header)